requirements.txt
fastapi==0.111.1
llama_index==0.10.58
llama_parse==0.4.9
markdownify==0.13.1
pandas==2.2.2
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
tiktoken==0.7.0
tqdm==4.66.4
uvicorn==0.30.3
weaviate_client==4.7.0
llama-cloud==0.0.6
llama-index==0.10.58
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.58
llama-index-embeddings-gemini==0.1.8
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.4
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-chroma==0.1.10
llama-index-vector-stores-weaviate==1.0.2
llama-index-readers-web==0.1.23
